Shomate polynomials were not properly implemented in PoKiTT, and this wasn't exposed in the tests due inadequate temperature range coverage.
Because reaction rates didn't support shomate, and because we have yet to see any combustion mechanisms that use shomate polynomials, we have elected to remove support for shomate altogether.  PoKiTT now throws an exception if the Cantera input files use shomate.

This commit also extends temperature bounds for the temperature and thermo tests.

This closes issue #5.